Rename package to organizerapi
