+++ OrganizerAPI/src/main/java/com/enfec/sb/organizerapi/OrganizerApiApplication.java
@@ -1,4 +1,4 @@
-package com.enfec.sb.deviceInfoapi;
+package com.enfec.sb.organizerapi;
 
 
 import org.springframework.boot.SpringApplication;
