+++ OrganizerAPI/src/main/java/com/enfec/sb/organizerapi/model/OrganizerTable.java
@@ -1,4 +1,4 @@
-package com.enfec.sb.deviceInfoapi.model;
+package com.enfec.sb.organizerapi.model;
 
 import org.springframework.stereotype.Component;
